Refactoring parameter factories
