--- iTopologyManager/parameterFactory/pom.xml
@@ -1,54 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ pom.xml
-  ~
-  ~ This work is free software; you can redistribute it and/or modify
-  ~ it under the terms of the GNU General Public License as published
-  ~ by the Free Software Foundation; either version 2 of the License,
-  ~ or (at your option) any later version.
-  ~
-  ~ This work is distributed in the hope that it will be useful, but
-  ~ WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-  ~ GNU General Public License for more details.
-  ~
-  ~ You should have received a copy of the GNU General Public License
-  ~ along with this program; if not, write to the Free Software
-  ~ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307
-  ~ USA
-  ~
-  ~ Copyright (c) 2010-2016 iTransformers Labs. All rights reserved.
-  -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>iTopologyManager</artifactId>
-        <groupId>net.itransformers</groupId>
-        <version>0.8</version><!--Change for new version-->
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>parameterFactory</artifactId>
-    <version>0.8</version><!--Change for new version-->
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-dependency-plugin</artifactId>
-                <version>2.2</version>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-surefire-plugin</artifactId>
-                <version>2.19.1</version>
-                <!--<configuration>-->
-                <!--<workingDirectory>${project.parent.parent.basedir}</workingDirectory>-->
-                <!--</configuration>-->
-            </plugin>
-
-        </plugins>
-    </build>
-
-</project>
