--- iTopologyManager/parameterFactory/src/main/java/net/itransformers/topologyviewer/parameterfactory/config/ObjectFactory.java
@@ -1,119 +1,0 @@
-/*
- * iTransformer is an open source tool able to discover and transform
- *  IP network infrastructures.
- *  Copyright (C) 2012  http://itransformers.net
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.4 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2012.01.18 at 05:12:39 AM EST 
-//
-
-
-package net.itransformers.topologyviewer.parameterfactory.config;
-
-import javax.xml.bind.JAXBElement;
-import javax.xml.bind.annotation.XmlElementDecl;
-import javax.xml.bind.annotation.XmlRegistry;
-import javax.xml.namespace.QName;
-
-
-/**
- * This object contains factory methods for each 
- * Java content interface and Java element interface 
- * generated in the net.itransformers.topologyviewer.parameterfactory.config package.
- * <p>An ObjectFactory allows you to programatically 
- * construct new instances of the Java representation 
- * for XML content. The Java representation of XML 
- * content can consist of schema derived interfaces 
- * and classes representing the binding of schema 
- * type definitions, element declarations and model 
- * groups.  Factory methods for each of these are 
- * provided in this class.
- * 
- */
-@XmlRegistry
-public class ObjectFactory {
-
-    private final static QName _ParamFactories_QNAME = new QName("", "param-factories");
-
-    /**
-     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: net.itransformers.topologyviewer.parameterfactory.config
-     * 
-     */
-    public ObjectFactory() {
-    }
-
-    /**
-     * Create an instance of {@link ParamFactoriesType }
-     * 
-     */
-    public ParamFactoriesType createParamFactoriesType() {
-        return new ParamFactoriesType();
-    }
-
-    /**
-     * Create an instance of {@link TypeType }
-     * 
-     */
-    public TypeType createTypeType() {
-        return new TypeType();
-    }
-
-    /**
-     * Create an instance of {@link ParamFactoryElementTypesType }
-     * 
-     */
-    public ParamFactoryElementTypesType createParamFactoryElementTypesType() {
-        return new ParamFactoryElementTypesType();
-    }
-
-    /**
-     * Create an instance of {@link ParamFactoryType }
-     * 
-     */
-    public ParamFactoryType createParamFactoryType() {
-        return new ParamFactoryType();
-    }
-
-    /**
-     * Create an instance of {@link ParamType }
-     * 
-     */
-    public ParamType createParamType() {
-        return new ParamType();
-    }
-
-    /**
-     * Create an instance of {@link ParamFactoryElementType }
-     * 
-     */
-    public ParamFactoryElementType createParamFactoryElementType() {
-        return new ParamFactoryElementType();
-    }
-
-    /**
-     * Create an instance of {@link JAXBElement }{@code <}{@link ParamFactoriesType }{@code >}}
-     * 
-     */
-    @XmlElementDecl(namespace = "", name = "param-factories")
-    public JAXBElement<ParamFactoriesType> createParamFactories(ParamFactoriesType value) {
-        return new JAXBElement<ParamFactoriesType>(_ParamFactories_QNAME, ParamFactoriesType.class, null, value);
-    }
-
-}
