--- iTopologyManager/parameterFactory/src/main/java/net/itransformers/topologyviewer/parameterfactory/config/ParamFactoriesType.java
@@ -1,125 +1,0 @@
-/*
- * ParamFactoriesType.java
- *
- * This work is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published
- * by the Free Software Foundation; either version 2 of the License,
- * or (at your option) any later version.
- *
- * This work is distributed in the hope that it will be useful, but
- * WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307
- * USA
- *
- * Copyright (c) 2010-2016 iTransformers Labs. All rights reserved.
- */
-
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.4 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2012.01.18 at 05:12:39 AM EST 
-//
-
-
-package net.itransformers.topologyviewer.parameterfactory.config;
-
-import java.util.ArrayList;
-import java.util.List;
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlType;
-
-
-/**
- * <p>Java class for param-factoriesType complex type.
- * 
- * <p>The following schema fragment specifies the expected content contained within this class.
- * 
- * <pre>
- * &lt;complexType name="param-factoriesType">
- *   &lt;complexContent>
- *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *       &lt;sequence>
- *         &lt;element name="param-factory-element-types" type="{}param-factory-element-typesType"/>
- *         &lt;element name="param-factory" type="{}param-factoryType" maxOccurs="unbounded" minOccurs="0"/>
- *       &lt;/sequence>
- *     &lt;/restriction>
- *   &lt;/complexContent>
- * &lt;/complexType>
- * </pre>
- * 
- * 
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "param-factoriesType", propOrder = {
-    "paramFactoryElementTypes",
-    "paramFactory"
-})
-public class ParamFactoriesType {
-
-    @XmlElement(name = "param-factory-element-types", required = true)
-    protected ParamFactoryElementTypesType paramFactoryElementTypes;
-    @XmlElement(name = "param-factory")
-    protected List<ParamFactoryType> paramFactory;
-
-    /**
-     * Gets the value of the paramFactoryElementTypes property.
-     * 
-     * @return
-     *     possible object is
-     *     {@link ParamFactoryElementTypesType }
-     *     
-     */
-    public ParamFactoryElementTypesType getParamFactoryElementTypes() {
-        return paramFactoryElementTypes;
-    }
-
-    /**
-     * Sets the value of the paramFactoryElementTypes property.
-     * 
-     * @param value
-     *     allowed object is
-     *     {@link ParamFactoryElementTypesType }
-     *     
-     */
-    public void setParamFactoryElementTypes(ParamFactoryElementTypesType value) {
-        this.paramFactoryElementTypes = value;
-    }
-
-    /**
-     * Gets the value of the paramFactory property.
-     * 
-     * <p>
-     * This accessor method returns a reference to the live list,
-     * not a snapshot. Therefore any modification you make to the
-     * returned list will be present inside the JAXB object.
-     * This is why there is not a <CODE>set</CODE> method for the paramFactory property.
-     * 
-     * <p>
-     * For example, to add a new item, do as follows:
-     * <pre>
-     *    getParamFactory().add(newItem);
-     * </pre>
-     * 
-     * 
-     * <p>
-     * Objects of the following type(s) are allowed in the list
-     * {@link ParamFactoryType }
-     * 
-     * 
-     */
-    public List<ParamFactoryType> getParamFactory() {
-        if (paramFactory == null) {
-            paramFactory = new ArrayList<ParamFactoryType>();
-        }
-        return this.paramFactory;
-    }
-
-}
