--- iTopologyManager/parameterFactory/src/main/resources/parameterFactory/conf/xml/serv-param-factory.xml
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ serv-param-factory.xml
-  ~
-  ~ This work is free software; you can redistribute it and/or modify
-  ~ it under the terms of the GNU General Public License as published
-  ~ by the Free Software Foundation; either version 2 of the License,
-  ~ or (at your option) any later version.
-  ~
-  ~ This work is distributed in the hope that it will be useful, but
-  ~ WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-  ~ GNU General Public License for more details.
-  ~
-  ~ You should have received a copy of the GNU General Public License
-  ~ along with this program; if not, write to the Free Software
-  ~ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307
-  ~ USA
-  ~
-  ~ Copyright (c) 2010-2016 iTransformers Labs. All rights reserved.
-  -->
-
-<param-factories xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:noNamespaceSchemaLocation="../xsd/param-factory.xsd">
-    <param-factory-element-types>
-        <type name="manual" class="com.parameterfactory.impl.GUIParamFactoryElement"/>
-        <type name="graphml" class="com.parameterfactory.impl.MapDerivedParamFactoryElement">
-            <param name="contextKey">graphml</param>
-        </type>
-        <type name="resource" class="com.parameterfactory.impl.MapDerivedParamFactoryElement">
-            <param name="contextKey">connection-params</param>
-        </type>
-        <type name="deviceData" class="com.parameterfactory.impl.XmlParamFactoryElement">
-            <param name="contextXmlFileNameKey">xmlFileName</param>
-        </type>
-    </param-factory-element-types>
-    <param-factory name="add">
-        <param-factory-element type="manual">
-            <param name="Service Name"/>
-            <param name="Service Type"/>
-            <param name="Service Locations"/>
-        </param-factory-element>
-        <param-factory-element type="resource">
-            <param name="username"/>
-            <param name="password"/>
-        </param-factory-element>
-    </param-factory>
-     <param-factory name="delete">
-        <param-factory-element type="manual">
-            <param name="Service Name"/>
-            <param name="Service Type"/>
-            <param name="Service Locations"/>
-        </param-factory-element>
-        <param-factory-element type="resource">
-            <param name="username"/>
-            <param name="password"/>
-        </param-factory-element>
-    </param-factory>
-    <param-factory name="modify">
-        <param-factory-element type="manual">
-            <param name="Service Name"/>
-            <param name="Service Type"/>
-            <param name="Service Locations"/>
-        </param-factory-element>
-        <param-factory-element type="resource">
-            <param name="username"/>
-            <param name="password"/>
-        </param-factory-element>
-    </param-factory>
-</param-factories>
