--- iTopologyManager/parameterFactory/src/main/resources/parameterFactory/conf/xsd/param-factory.xsd
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ param-factory.xsd
-  ~
-  ~ This work is free software; you can redistribute it and/or modify
-  ~ it under the terms of the GNU General Public License as published
-  ~ by the Free Software Foundation; either version 2 of the License,
-  ~ or (at your option) any later version.
-  ~
-  ~ This work is distributed in the hope that it will be useful, but
-  ~ WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-  ~ GNU General Public License for more details.
-  ~
-  ~ You should have received a copy of the GNU General Public License
-  ~ along with this program; if not, write to the Free Software
-  ~ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307
-  ~ USA
-  ~
-  ~ Copyright (c) 2010-2016 iTransformers Labs. All rights reserved.
-  -->
-
-<xs:schema attributeFormDefault="unqualified" elementFormDefault="qualified" xmlns:xs="http://www.w3.org/2001/XMLSchema">
-  <xs:element name="param-factories" type="param-factoriesType"/>
-  <xs:complexType name="typeType">
-    <xs:sequence>
-      <xs:element type="paramType" name="param" maxOccurs="unbounded" minOccurs="0"/>
-    </xs:sequence>
-    <xs:attribute type="xs:string" name="name" use="required"/>
-    <xs:attribute type="xs:string" name="class" use="required"/>
-  </xs:complexType>
-  <xs:complexType name="param-factory-element-typesType">
-    <xs:sequence>
-      <xs:element type="typeType" name="type" maxOccurs="unbounded" minOccurs="0"/>
-    </xs:sequence>
-  </xs:complexType>
-  <xs:complexType name="param-factoriesType">
-    <xs:sequence>
-      <xs:element type="param-factory-element-typesType" name="param-factory-element-types"/>
-      <xs:element type="param-factoryType" name="param-factory" maxOccurs="unbounded" minOccurs="0"/>
-    </xs:sequence>
-  </xs:complexType>
-  <xs:complexType name="param-factoryType">
-    <xs:sequence>
-      <xs:element type="param-factory-elementType" name="param-factory-element" maxOccurs="unbounded" minOccurs="0"/>
-    </xs:sequence>
-    <xs:attribute type="xs:string" name="name" use="optional"/>
-  </xs:complexType>
-  <xs:complexType name="param-factory-elementType">
-    <xs:sequence>
-      <xs:element type="paramType" name="param" maxOccurs="unbounded" minOccurs="0"/>
-    </xs:sequence>
-    <xs:attribute type="xs:string" name="type" use="required"/>
-  </xs:complexType>
-  <xs:complexType name="paramType">
-    <xs:simpleContent>
-      <xs:extension base="xs:string">
-        <xs:attribute type="xs:string" name="name" use="required"/>
-      </xs:extension>
-    </xs:simpleContent>
-  </xs:complexType>
-</xs:schema>
