+++ parameterFactory/parameterFactoryAPI/pom.xml
@@ -1,0 +1,52 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>parameterFactory</artifactId>
+        <groupId>net.itransformers</groupId>
+        <version>0.8</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <groupId>net.itransformers</groupId>
+    <artifactId>parameterFactoryAPI</artifactId>
+
+     <build>
+         <plugins>
+             <plugin>
+                 <groupId>org.apache.maven.plugins</groupId>
+                 <artifactId>maven-compiler-plugin</artifactId>
+                 <configuration>
+                     <source>1.8</source>
+                     <target>1.8</target>
+                 </configuration>
+             </plugin>
+
+             <plugin>
+                 <groupId>org.codehaus.mojo</groupId>
+                 <artifactId>jaxb2-maven-plugin</artifactId>
+                 <version>1.6</version>
+                 <executions>
+                     <execution>
+                         <id>xjc-model</id>
+                         <goals>
+                             <goal>xjc</goal>
+                         </goals>
+                         <phase>generate-sources</phase>
+                         <configuration>
+                             <schemaDirectory>${basedir}/src/main/resources/conf/xsd</schemaDirectory>
+                             <!-- The package of your generated sources -->
+                             <packageName>net.itransformers.parameterfactoryapi.model</packageName>
+                             <clearOutputDir>false</clearOutputDir>
+                             <schemaFiles>param-factory.xsd</schemaFiles>
+                             <outputDirectory>${basedir}/src/main/java</outputDirectory>
+                         </configuration>
+                     </execution>
+
+                 </executions>
+
+             </plugin>
+         </plugins>
+     </build>
+</project>
