+++ parameterFactory/parameterFactoryAPI/src/main/java/net/itransformers/parameterfactoryapi/ParameterFactoryException.java
@@ -1,0 +1,15 @@
+package net.itransformers.parameterfactoryapi;
+
+/**
+ * Created by niau on 1/12/17.
+ */
+public class ParameterFactoryException extends RuntimeException {
+    public ParameterFactoryException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public ParameterFactoryException(String message) {
+        super(message);
+    }
+
+}
