+++ parameterFactory/parameterFactoryAPI/src/main/java/net/itransformers/parameterfactoryapi/ParameterFactoryManagerFactory.java
@@ -1,0 +1,10 @@
+package net.itransformers.parameterfactoryapi;
+
+import java.util.Map;
+
+/**
+ * Created by niau on 1/9/17.
+ */
+public interface ParameterFactoryManagerFactory  {
+    ParameterFactoryManger createParameterFactorysManager(Map<String, String> properties);
+}
