+++ parameterFactory/parameterFactoryAPI/src/main/java/net/itransformers/parameterfactoryapi/model/ObjectFactory.java
@@ -1,0 +1,118 @@
+//
+// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.7 
+// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
+// Any modifications to this file will be lost upon recompilation of the source schema. 
+// Generated on: 2017.01.18 at 03:34:32 AM EET 
+//
+
+
+package net.itransformers.parameterfactoryapi.model;
+
+import javax.xml.bind.JAXBElement;
+import javax.xml.bind.annotation.XmlElementDecl;
+import javax.xml.bind.annotation.XmlRegistry;
+import javax.xml.namespace.QName;
+
+
+/**
+ * This object contains factory methods for each 
+ * Java content interface and Java element interface 
+ * generated in the net.itransformers.parameterfactoryapi.model package. 
+ * <p>An ObjectFactory allows you to programatically 
+ * construct new instances of the Java representation 
+ * for XML content. The Java representation of XML 
+ * content can consist of schema derived interfaces 
+ * and classes representing the binding of schema 
+ * type definitions, element declarations and model 
+ * groups.  Factory methods for each of these are 
+ * provided in this class.
+ * 
+ */
+@XmlRegistry
+public class ObjectFactory {
+
+    private final static QName _ParamFactories_QNAME = new QName("", "param-factories");
+    private final static QName _ParamFactoriesConfig_QNAME = new QName("", "param-factories-config");
+
+    /**
+     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: net.itransformers.parameterfactoryapi.model
+     * 
+     */
+    public ObjectFactory() {
+    }
+
+    /**
+     * Create an instance of {@link ParamFactoriesType }
+     * 
+     */
+    public ParamFactoriesType createParamFactoriesType() {
+        return new ParamFactoriesType();
+    }
+
+    /**
+     * Create an instance of {@link ParamFactoriesConfigType }
+     * 
+     */
+    public ParamFactoriesConfigType createParamFactoriesConfigType() {
+        return new ParamFactoriesConfigType();
+    }
+
+    /**
+     * Create an instance of {@link ParamType }
+     * 
+     */
+    public ParamType createParamType() {
+        return new ParamType();
+    }
+
+    /**
+     * Create an instance of {@link TypeType }
+     * 
+     */
+    public TypeType createTypeType() {
+        return new TypeType();
+    }
+
+    /**
+     * Create an instance of {@link ParamFactoryElementType }
+     * 
+     */
+    public ParamFactoryElementType createParamFactoryElementType() {
+        return new ParamFactoryElementType();
+    }
+
+    /**
+     * Create an instance of {@link ParamFactoryType }
+     * 
+     */
+    public ParamFactoryType createParamFactoryType() {
+        return new ParamFactoryType();
+    }
+
+    /**
+     * Create an instance of {@link ParamFactoryElementTypesType }
+     * 
+     */
+    public ParamFactoryElementTypesType createParamFactoryElementTypesType() {
+        return new ParamFactoryElementTypesType();
+    }
+
+    /**
+     * Create an instance of {@link JAXBElement }{@code <}{@link ParamFactoriesType }{@code >}}
+     * 
+     */
+    @XmlElementDecl(namespace = "", name = "param-factories")
+    public JAXBElement<ParamFactoriesType> createParamFactories(ParamFactoriesType value) {
+        return new JAXBElement<ParamFactoriesType>(_ParamFactories_QNAME, ParamFactoriesType.class, null, value);
+    }
+
+    /**
+     * Create an instance of {@link JAXBElement }{@code <}{@link ParamFactoriesConfigType }{@code >}}
+     * 
+     */
+    @XmlElementDecl(namespace = "", name = "param-factories-config")
+    public JAXBElement<ParamFactoriesConfigType> createParamFactoriesConfig(ParamFactoriesConfigType value) {
+        return new JAXBElement<ParamFactoriesConfigType>(_ParamFactoriesConfig_QNAME, ParamFactoriesConfigType.class, null, value);
+    }
+
+}
