+++ parameterFactory/parameterFactoryAPI/src/main/java/net/itransformers/parameterfactoryapi/util/JaxbMarshalar.java
@@ -1,0 +1,49 @@
+/*
+ * JaxbMarshalar.java
+ *
+ * This work is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published
+ * by the Free Software Foundation; either version 2 of the License,
+ * or (at your option) any later version.
+ *
+ * This work is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307
+ * USA
+ *
+ * Copyright (c) 2010-2016 iTransformers Labs. All rights reserved.
+ */
+
+package net.itransformers.parameterfactoryapi.util;
+
+import javax.xml.bind.*;
+import javax.xml.namespace.QName;
+import javax.xml.transform.stream.StreamSource;
+import java.io.InputStream;
+import java.io.OutputStream;
+
+public class JaxbMarshalar {
+    public static <T> void marshal(T obj, OutputStream outputStream, String rootTagName) throws JAXBException {
+        Class docClass = obj.getClass();
+        String packageName = docClass.getPackage().getName();
+        JAXBContext jc = JAXBContext.newInstance(packageName);
+        Marshaller m = jc.createMarshaller();
+        m.setProperty(Marshaller.JAXB_FORMATTED_OUTPUT, Boolean.TRUE);
+        m.marshal( new JAXBElement<T>(new QName("",rootTagName), docClass, obj), outputStream );
+    }
+
+    @SuppressWarnings("unchecked")
+    public static <T> T unmarshal( Class<T> docClass, InputStream inputStream ) throws JAXBException {
+        String packageName = docClass.getPackage().getName();
+        JAXBContext jc = JAXBContext.newInstance(packageName);
+        Unmarshaller u = jc.createUnmarshaller();
+        JAXBElement<T> doc = (JAXBElement<T>)u.unmarshal( new StreamSource(inputStream), docClass );
+        return doc.getValue();
+    }
+
+}
